[FIX] fix guard

+++ app/Http/Controllers/Admin/ContentController.php
@@ -117,9 +117,13 @@ class ContentController extends Controller
      */
     public function edit(Content $konten)
     {
-        return view("contents.edit")->with([
-            'konten' => $konten
-        ]);
+        if($konten->user_id == Auth::user()->id )
+        {
+            return view("contents.edit")->with([
+                'konten' => $konten
+            ]);
+        }
+        abort(403, "Anda tidak bisa memodifikasi konten yang bukan buatan Anda");
     }
 
     /**
@@ -127,72 +131,76 @@ class ContentController extends Controller
      */
     public function update(Request $request, Content $konten)
     {
-        $request->validate([
-            "judul" => "required|unique:contents,judul," . $konten->id,
-            "user_id" => "numeric|exists:users,id",
-            "kategori" => "required|in:Proker,Artikel",
-            "konten" => "required",
-            "gambar_andalan_konten" => "max:1024|mimes:webp",
-            "galeri_konten" => "nullable|max:5000|mimes:zip",
-        ]);
-
-        DB::beginTransaction();
-        try {
-            $konten->update([
-                "judul" => $request->judul,
-                "slug" => Str::slug($request->judul),
-                "user_id" => Auth::user()->id,
-                "kategori" => $request->kategori,
-                "konten" => $request->konten
+        if($konten->user_id == Auth::user()->id )
+        {
+            $request->validate([
+                "judul" => "required|unique:contents,judul," . $konten->id,
+                "user_id" => "numeric|exists:users,id",
+                "kategori" => "required|in:Proker,Artikel",
+                "konten" => "required",
+                "gambar_andalan_konten" => "max:1024|mimes:webp",
+                "galeri_konten" => "nullable|max:5000|mimes:zip",
             ]);
-
-            if($request->hasFile("gambar_andalan_konten"))
-            {
-                $konten->clearMediaCollection("gambar_andalan_konten");
-                $konten->addMediaFromRequest("gambar_andalan_konten")->toMediaCollection("gambar_andalan_konten");
-            }
-
-            // check if gallery is not null
-            if($request->hasFile("galeri_konten")) {
-                // get zip file
-                $zip_file = $request->file("galeri_konten");
-
-                // extract it
-                $zip = new ZipArchive;
-                $zip->open($zip_file);
-                $zip->extractTo(public_path()."/zip/");
-                $zip->close();
-
-                // get all gallery
-                $galleries = File::allFiles(public_path()."/zip/");
-
-                $is_valid = true;
-                // loop galleries to store into media
-                foreach($galleries as $gallery)
+    
+            DB::beginTransaction();
+            try {
+                $konten->update([
+                    "judul" => $request->judul,
+                    "slug" => Str::slug($request->judul),
+                    "user_id" => Auth::user()->id,
+                    "kategori" => $request->kategori,
+                    "konten" => $request->konten
+                ]);
+    
+                if($request->hasFile("gambar_andalan_konten"))
                 {
-                    $extension = $gallery->getExtension();
-                    if($extension == "webp") {
-                        $konten->addMedia($gallery->getRealPath())->toMediaCollection("galeri_konten");
-                        File::delete($gallery->getRealPath());
-                    } else {
-                        File::delete($gallery->getRealPath());
-                        $is_valid = false;
+                    $konten->clearMediaCollection("gambar_andalan_konten");
+                    $konten->addMediaFromRequest("gambar_andalan_konten")->toMediaCollection("gambar_andalan_konten");
+                }
+    
+                // check if gallery is not null
+                if($request->hasFile("galeri_konten")) {
+                    // get zip file
+                    $zip_file = $request->file("galeri_konten");
+    
+                    // extract it
+                    $zip = new ZipArchive;
+                    $zip->open($zip_file);
+                    $zip->extractTo(public_path()."/zip/");
+                    $zip->close();
+    
+                    // get all gallery
+                    $galleries = File::allFiles(public_path()."/zip/");
+    
+                    $is_valid = true;
+                    // loop galleries to store into media
+                    foreach($galleries as $gallery)
+                    {
+                        $extension = $gallery->getExtension();
+                        if($extension == "webp") {
+                            $konten->addMedia($gallery->getRealPath())->toMediaCollection("galeri_konten");
+                            File::delete($gallery->getRealPath());
+                        } else {
+                            File::delete($gallery->getRealPath());
+                            $is_valid = false;
+                        }
+                    }
+                    if(!$is_valid){
+                        DB::rollBack();
+                        session()->flash('error', 'File yang didalam zip harus berupa webp');
+                        File::delete($zip_file);
+                        return redirect()->back();
                     }
                 }
-                if(!$is_valid){
-                    DB::rollBack();
-                    session()->flash('error', 'File yang didalam zip harus berupa webp');
-                    File::delete($zip_file);
-                    return redirect()->back();
-                }
+                DB::commit();
+                return redirect()->route("admin.konten.index")->with("success", "Data berhasil diubah");
+            } catch (Exception $e) {
+                //throw $th;
+                DB::rollBack();
+                return redirect()->back()->with("error", $e->getMessage());
             }
-            DB::commit();
-            return redirect()->route("admin.konten.index")->with("success", "Data berhasil diubah");
-        } catch (Exception $e) {
-            //throw $th;
-            DB::rollBack();
-            return redirect()->back()->with("error", $e->getMessage());
         }
+        abort(403, "Anda tidak bisa memodifikasi konten yang bukan buatan Anda");
     }
 
     /**
@@ -200,25 +208,34 @@ class ContentController extends Controller
      */
     public function destroy(Content $konten)
     {
-        // Delete Viewers
-        ContentView::whereContentId($konten->id)->delete();
-        
-        // Delete Comment
-        Comment::whereContentId($konten->id)->delete();
+        if($konten->user_id == Auth::user()->id )
+        {
+            // Delete Viewers
+            ContentView::whereContentId($konten->id)->delete();
 
-        // Delete media and data
-        $konten->clearMediaCollection("gambar_andalan_konten");
-        $konten->clearMediaCollection("galeri_konten");
-        $konten->delete();
-        return redirect()->back()->with("success", "Data berhasil dihapus");
+            // Delete Comment
+            Comment::whereContentId($konten->id)->delete();
+    
+            // Delete media and data
+            $konten->clearMediaCollection("gambar_andalan_konten");
+            $konten->clearMediaCollection("galeri_konten");
+            $konten->delete();
+            return redirect()->back()->with("success", "Data berhasil dihapus");
+        }
+        abort(403, "Anda tidak bisa memodifikasi konten yang bukan buatan Anda");
+        
     }
 
     public function comment(Content $konten)
     {
-        $comments = Comment::with("Member")->whereContentId($konten->id)->get();
-        return view("contents.comments")->with([
-            "comments" => $comments
-        ]);
+        if($konten->user_id == Auth::user()->id )
+        {
+            $comments = Comment::with("Member")->whereContentId($konten->id)->get();
+            return view("contents.comments")->with([
+                "comments" => $comments
+            ]);
+        }
+        abort(403, "Anda tidak bisa memodifikasi konten yang bukan buatan Anda");
     }
 
     public function destroy_comment(Comment $komentar)
